requirements.txt
# Core dependencies
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Oura API
requests>=2.31.0

# Azure OpenAI
openai>=2.0.0

# Telegram
python-telegram-bot>=21.0

# Scheduling
APScheduler>=3.10.4

# Utilities
python-dateutil>=2.8.2
